Refactoring, needs more comments
